Split up more components in BestPractices and MealTips
