+++ src/components/Training/module-2/components/index.js
@@ -1,4 +1,6 @@
 import { SimpleStep } from './SimpleStep'
 import { FiveSteps } from './FiveSteps'
+import { BestPractices } from './BestPractices'
+import { MealTips } from './MealTips'
 
-export { SimpleStep, FiveSteps }
+export { SimpleStep, FiveSteps, BestPractices, MealTips }
